requirements.txt
Flask
flask_cors
requests
msal
gunicorn
flask_talisman
pytz
python_dotenv